+++ src/Spieler.cpp
@@ -3,7 +3,7 @@
 
 #include "Spieler.hpp"
 
-Spieler::Spieler(std::string filename, PlayerHandler* mplayerhandler, SchussHandler* mschusshandler, float x, float y)
+Spieler::Spieler(std::string filename, PlayerHandler* mplayerhandler, SchussHandler* mschusshandler, double x, double y)
 : Object(filename,SPIELER, x, y), mPlayerHandler(mplayerhandler), mSchussHandler(mschusshandler), SchussCoolDown(0)
 {
 }
